% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{parse_idxstats}
\alias{parse_idxstats}
\title{Parse idxstats file}
\usage{
parse_idxstats(path)
}
\arguments{
\item{path}{to idxstats file produced by \verb{samtools idxstats}}
}
\value{
data.frame with columns "reference","length", "n_mapped" and "n_unmapped".
}
\description{
parse idxstats file.
}
\examples{
path = system.file(package="mitochondrir", "idxstats")
parse_idxstats(path)
}
