% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_mtdna}
\alias{calculate_mtdna}
\title{calculate mitochondrial dna burden}
\usage{
calculate_mtdna(
  mitochondrial_depth,
  autosome_depth,
  ploidy = 2,
  per_cell = TRUE
)
}
\arguments{
\item{mitochondrial_depth}{depth of coverage of the mitochondrial genome}

\item{autosome_depth}{depth of coverage of the autosomal genome (do not normalise for ploidy yet! this function performs the normalisation for you)}

\item{ploidy}{autosomal ploidy. By default assumes cells are diploid (ploidy = 2)}

\item{per_cell}{return mitochondrial genomes per cell. If FALSE, return mitochondrial genomes per autosome (single copy)}
}
\value{
Number of mitochondrial genomes per cell. If \code{per_cell=false} returns number of mitochondrial genomes per single copy of the autosome.
}
\description{
calculate mitochondrial dna burden
}
\examples{
calculate_mtdna(
  mitochondrial_depth = 100,
  autosome_depth = 10,
  ploidy = 2
)
}
