% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_mtdna_tumour}
\alias{calculate_mtdna_tumour}
\title{Calculate typical number of mitochondria per cell}
\usage{
calculate_mtdna_tumour(
  mitochondrial_depth,
  autosome_depth,
  background_ploidy = 2,
  tumour_ploidy = 2,
  tumour_purity = 1,
  per_cell = TRUE
)
}
\arguments{
\item{mitochondrial_depth}{depth of coverage of the mitochondrial genome}

\item{autosome_depth}{depth of coverage of the autosomal genome (do not normalise for ploidy yet! this function performs the normalisation for you)}

\item{background_ploidy}{expected ploidy of healthy cells in bulk wgs sample (default: 2 for diploid background)}

\item{tumour_ploidy}{median ploidy of tumour cells in bulk sample (default: 2 for diploid tumours)}

\item{tumour_purity}{proportion of total cells in sample that are tumour cells (must be between 0 and 1). used to correct for tumour ploidy.}

\item{per_cell}{return mitochondrial genomes per cell. If FALSE, return mitochondrial genomes per autosome (single copy)}
}
\value{
number of mitochondrial genomes per cell. if \code{per_cell=false} returns number of mitochondrial genomes per single copy of the autosome.
}
\description{
from a bulk whole-genome tumour sample, calculate number of mitochondrial genomes per cell.
this function includes options to correct for tumour ploidy and purity,
but by default will assume all cells in sample are diploid
}
\examples{
calculate_mtdna_tumour(
  mitochondrial_depth = 100,
  autosome_depth = 10,
  background_ploidy = 2,
  tumour_ploidy = 2,
  tumour_purity = 1
)
}
