% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{mtdna_from_idxstats}
\alias{mtdna_from_idxstats}
\title{Calculate mtdna burden from idxstats file}
\usage{
mtdna_from_idxstats(
  path,
  background_ploidy = 2,
  tumour_ploidy = 2,
  tumour_purity = 1,
  mtname = "chrM",
  autosomes = paste0("chr", 1:22),
  per_cell = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to idxstats file produced by \verb{samtools idxstats <bam>}}

\item{background_ploidy}{expected ploidy of healthy cells in bulk wgs sample (default: 2 for diploid background)}

\item{tumour_ploidy}{median ploidy of tumour cells in bulk sample (default: 2 for diploid tumours)}

\item{tumour_purity}{proportion of total cells in sample that are tumour cells (must be between 0 and 1). used to correct for tumour ploidy.}

\item{mtname}{name of mitochondrial genome in reference genome}

\item{autosomes}{name of autosomal sequences in reference genome}

\item{per_cell}{return mitochondrial genomes per cell. If FALSE, return mitochondrial genomes per autosome (single copy)}

\item{verbose}{verbose mode}
}
\value{
number of mitochondrial genomes per cell. if \code{per_cell=FALSE} returns number of mitochondrial genomes per single copy of the autosome.
}
\description{
Calculate mtdna burden from idxstats file
}
\examples{
path = system.file(package="mitochondrir", "idxstats")
parse_idxstats(path)
}
