% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{mtdna_from_idxstats}
\alias{mtdna_from_idxstats}
\title{Calculate mtdna burden from idxstats file}
\usage{
mtdna_from_idxstats(
  path,
  background_ploidy = 2,
  tumour_ploidy = 2,
  tumour_purity = 1,
  read_length = 150,
  mtname = "chrm",
  autosomes = paste0("chr", 1:22),
  per_cell = TRUE,
  verbose = true
)
}
\arguments{
\item{path}{Path to idxstats file produced by \verb{samtools idxstats <bam>}}

\item{background_ploidy}{expected ploidy of healthy cells in bulk wgs sample (default: 2 for diploid background)}

\item{tumour_ploidy}{median ploidy of tumour cells in bulk sample (default: 2 for diploid tumours)}

\item{tumour_purity}{proportion of total cells in sample that are tumour cells (must be between 0 and 1). used to correct for tumour ploidy.}

\item{read_length}{Length of each read. Used to estimate depth but if not known, feel free to leave at default.
It will change the absolute values of depths of coverage but not affect the calculation of mitochondrial burden.}

\item{mtname}{name of mitochondrial genome in reference genome}

\item{autosomes}{name of autosomal sequences in reference genome}

\item{per_cell}{return mitochondrial genomes per cell. If FALSE, return mitochondrial genomes per autosome (single copy)}

\item{verbose}{verbose mode}
}
\value{
number of mitochondrial genomes per cell. if \code{per_cell=false} returns number of mitochondrial genomes per single copy of the autosome.
}
\description{
Calculate mtdna burden from idxstats file
}
\examples{
path = system.file(package="mitochondrir", "idxstats")
parse_idxstats(path)
}
